Changed mpi datatype offset type to ptrdiff_t from long

+++ source/mpid.d
@@ -185,7 +185,7 @@ template toMPIType(T)
 
 						Datatype newDataType;
 
-						long[] offsets;
+						ptrdiff_t[] offsets;
 						Datatype[] dataTypes;
 						int[] blocklengths;
